Change color 10 from pink to green
This color is in the "bright green" slot so I prefer a green (darker
since this is the light scheme) to be here, this looks much better in my
daily use compared to before, when it seemed like pink and purple were
overused. Also lines up with what I need in the terminal 16 color set,
where the pink was very confusing when it shows up in diffs instead of
green.